chore(test): add pytest options in root
To help ensure there is a unified experience when running any/all tests,
we need to define this in the root directory. The rest of the
example-specific `conftest.py` can remain in `examples/conftest.py`.

Running `pytest` in the root directory will now correctly execute _all_
tests, including the examples.
